move tests to e2e
