+++ javascript/sdk/.env.test
@@ -1,0 +1,1 @@
+VITE_PRELUDE_SERVICE_URL="https://detect.dev.prelude.org"
