+++ javascript/sdk/vite.config.ts
@@ -1,3 +1,5 @@
+/// <reference types="vitest" />
+
 import { resolve } from "path";
 import { defineConfig } from "vite";
 
@@ -9,5 +11,8 @@ export default defineConfig({
       fileName: "sdk",
     },
     rollupOptions: {},
+  },
+  test: {
+    testTimeout: 10_000,
   },
 });
